Don't expand tabs in KateDocument::editInsertText().
This fixes Undo, which expanded tabs when "Insert spaces instead of
tabulators" is enabled.

It also affects the behavior of transpose() if one of the transposed
characters is a tab. The old behavior is that the tab is replaced by
spaces, whereas the new behavior keeps the tab even if "Insert spaces instead
of tabulators" is enabled.
CCBUG: 113135

svn path=/trunk/KDE/kdelibs/; revision=525138