Consolidate Backspace's unindentation.
When the option "Backspace indents" is on, it unindents different from
Ctrl-Shift-I. Example:

f() {
    indentwidth_is_4();
    g(foo,
      bar);
        here();               // (*)
}

Put the cursor in front of here(). Now C-S-I unindents so that the line aligns
with indentwidth_is_4(). But Backspace unindents the line to align with bar,
which does not quite match the expectations. This fixes the problem by
replacing the custom unindentation implementation in KateDocument::backspace()
by a call to indent().

svn path=/trunk/KDE/kdelibs/; revision=525140