Adding missing CrystalSVG icon: cr16-action-tab_new_bg Adding new icon: cr16-action-tab_remove_other Updating two icons so the styles match: cr16-action-tab_remove, cr16-action-tab_new
Adding inherits to HiColor so that it is possible to use it as a theme.  If this doesn't work, it is due to bugs in the code.  We need this IAC since we do not have a complete HiColor theme.

svn path=/branches/KDE/3.5/kdelibs/; revision=525834

The following changes were in SVN, but were removed from git:

M	pics/crystalsvg/cr16-action-tab_new.png
A	pics/crystalsvg/cr16-action-tab_new_bg.png
M	pics/crystalsvg/cr16-action-tab_remove.png
A	pics/crystalsvg/cr16-action-tab_remove_other.png
M	pics/hicolor/index.theme