Even more refactoring of RMB. It now even shows a context menu. Currently hardcoded to the first bookmark, but otherwise it seems to work.
And removes some cruft of an apparently hidden feature. A feature without
a gui, and no google hit for the configuration option.

The feature removed the "Add Bookmark Here" and similar entries from the
menu and added a special submenu instead.

If you want to see it in action, try adding:
QuickActionSubmenu=true
to kbookmarkrc.

Well, that feature is gone.


svn path=/trunk/KDE/kdelibs/; revision=537286