Separate mimetype and servicetype factories. (Necessary for kdecore-kio split but also to change the mimetype implementation later in order to follow the XDG standard).
Clean up the trader code further, which also makes it twice faster since the
ksycoca lookup was being done twice (when a profile entry is available).

Remove support for reading .kdelnk files

KMimeType::diagnoseFileName (an API added by Waldo in an unrelated commit ;) has been replaced with KMimeType::extractKnownExtension which returns the extension only, not the pattern, which is what all the kde code using that method was using it for.
Also added KMimeType::isDefault() for more convenient checking of the result of findByURL.

I'll check tomorrow but I think we're ready for the kdecore move now.

svn path=/trunk/KDE/kdelibs/; revision=543819