Make wordLeft() and wordRight() behavior symmetric.
Consider this line:

  if (num >= 0) num--; else return;

The points where the cursor stops are (indicated by |):

  |if |(|num |>= |0|) |num|--; |else |return|;|

with both Ctrl-Left and Ctrl-Right.

This behavior _is_ different from the word move operations of some other
edit widgets (like, for example, KMail's composer), but it is more
appropriate for a programmer's editor, because it stops at all punctuation.
BUG: 56337

svn path=/branches/KDE/3.5/kdelibs/; revision=550059