* indent after '{'  * unintend if you type a '{'  * unindent if you type '}' and there is a corresponding '{'
If you type a '}', the script searches the corresponding '{'.
Use case: If you have your source code wrapped by a namespace {...}, then
it is likely to have severyl thousand lines between the { and the colosing }.
My test: If you type the '}' 3000 lines after the opening '{', the script
takes 1000ms (1 second!) to indent the line.

1000ms is a delay you notice, not that nice... --> todo: find workarounds

svn path=/trunk/KDE/kdelibs/; revision=570006