Object bindings no longer use void* to store the data, we use a real pointer, and a pointer_cast method to access it.  Because we do not use RTTI (maybe not sure though) by default there are some hacks in the painter bindings to get around the use of multiple inheritence in QWidget.  It seems because QRegion is not bound we throw an exception in the events bindings.  I removed the line for now, but we need a better way to deal with missing values inside of C++ methods.
I also tried to clean up some of the throwError calls so that it makes 
the error messages a bit easier to read.

Fixed the scribble demo, so now it at least works.

Renamed the kjs embed console app so it is not named something stupid 
like console.

CCMAIL rich@kde.org
CCMAIL mbroadst@gmail.com


svn path=/trunk/KDE/kdelibs/; revision=572711