Fix incorrect 'endl' usage introduced by #100188 fix.
The fix for bug #100188 replaces 'ws' by 'endl' when reading the lock
file. However, 'endl' isn't supposed to be used for reading and it was
trying to write a newline to the lockfile, causing "QFile::writeBlock:
Write operation not permitted" messages. Luckly it wasn't corrupting
the lockfile contents because it was opened read-only.

CCBUG: 100188


svn path=/branches/KDE/3.5/kdelibs/; revision=574367