Patch from Modestas Vainius <geromanas mailas.com> to fix Bug 122020 (kdesvnd action menu issues) He said, and I agree:
Do not time out when waiting for the reply for the X-KDE-GetActionMenu DCOP
call. DCOP protocol is synchronouos by design and a reply for the timed out
call (i.e. garbage) still arrives and might be falsely returned as a valid
reply for the subsequent DCOP call on the same connection. The application
currently known to be badly hurt by this is kdesvn and its konqueror service
menu integration. Its X-KDE-GetActionMenu might take longer than current 100
milisecond timeout to complete. As a result then, a late reply will "pollute" a
main DCOP connection of the client application (e.g. konqueror) causing further
DCOP failures.

svn path=/branches/KDE/3.5/kdelibs/; revision=599200