Fix bug in the infinite loop (while(true)) introduced in r588299:
When the text content passed to createMessageBox has the preffered width of
zero (either because the string isEmpty() or is e.g. "<qt></qt>" etc) I got
this: new_height and pref_height are always 21. used_width is always 0.

In Qt3 QTextDocument::widthUsed added a border_tolerance=2 to the widthUsed,
so seems to never be zero there.

Now, the Kate app starts again...

CCMAIL: tokoe@kde.org

svn path=/trunk/KDE/kdelibs/; revision=605665