Make vertical-align text-top/bottom more CSS 2.1 compliant: we are looking for the parent's font or font strut's *content area* (as per 10.6.1) so do not use its baselinePosition.
Add the missing half-leading for both text-top and text-bottom
(mozilla/css_1/vertical_align_texttop.html): that's also part of
content area.

Using baselinePosition on parent was even more bogus because of the
major semantic differences in CSS between baseline in table
context and other baselines. A mess that was the origin of #103039

BUG: 103039


svn path=/trunk/KDE/kdelibs/; revision=620738