Fixed 6-years old bug 20532: Undoing a copy operation might delete a modified file causing loss of data. A confirmation message box is shown when undoing a copy after modifying the file [this relies on copy preserving the mtime though, which might not be true for all protocols but we'll have to fix that].
The less severe case of move, modify, undo (no data loss, but the modified file is moved back) can't be fixed in all cases, since moving a whole directory can be done by a simple rename() call, so we have no information on all the files inside the directory. For single files it could be done though, but that's not implemented by this.

BUG: 20532

svn path=/trunk/KDE/kdelibs/; revision=626835