Start to fix internal highlighting code - it used different path to the external highlighting.  Now there is still a delay to first draw issue, but I'm not sure what's causing that yet.
Change the search highlighting code to use an all-encompassing range.
Not sure why this was required (or even if it is any more).  Might be
related to the above issue.

Seems that we have to watch all highlighted ranges rather than just
dynamically highlighted ones (much easier, and hopefully not really any
slower)

Also add a few signals to the code completion + implement.

CCMAIL:anders@alweb.dk

svn path=/trunk/KDE/kdelibs/; revision=639940