Move KConfigGroup's QtGui requirements into kdeui. Use the same hack that QVariant uses to access Gui types.
Accordingly, move the tests in kconfigtest that dealt with GUI types
into kconfigguitest.

This change adds a symbol to kdecore (_k_internal_KConfigGroupGui)
which means you have to rebuild kdecore if you rebuild kdeui (CMake
will do that automatically for you). Since the symbol is not public,
this change is binary- and source-compatible (though it's not
behaviour-compatible if you're using KConfig with QColor or QFont
without linking to kdeui).

svn path=/trunk/KDE/kdelibs/; revision=642959