Merge in work done thus far from the blizzard branch:
Basically, create a locals array in ActivationImp, and try to statically bind 
access to it as much as possible, providing a fallback symbolic path for when needed.
Also use the scopeing information to know to optimistically (unless eval of a var statement 
injects a local) skip local scope for lookup of variables not in there.

Also fixes some bugs in funcDecl handling....

Speed up results:

Richmoore's MD5:     7078ms to 5351ms (takes 75.6% of original runtime)
Richmoore's Cordic:  6900ms to 4210ms (takes 61  % of original runtime)
iBench JS:           1.48s to 1.29s   (takes 87  % of original runtime)

(It's stable, but needs a few more cleanups. Testcases coming soon on a non-monday..)


svn path=/trunk/KDE/kdelibs/; revision=647005