- fine-tune knewstuff2_test do not download when we don't want download - fix the i18n-id mechanism: if assumption holds that languages only get added, it works now
Interestingly, if Hotstuff is optimised to only deliver in the most-wanted user language,
the identification will still fail. There's no way around that. Very bad - do we really need
a server-generated id?

Also of interest is that KIO doesn't use the same language preference which is used in
Konqueror. Will have to find the reason.


svn path=/branches/work/knewstuff2/knewstuff2/; revision=648116