Ok, the big mount/umount/eject reorganization.
Now everything regarding eject is on OpticalDrive as it should be.
The mount/umount features are now more abstract (more portable, and that 
can do more than simply mount, think crypto volume, or unmount, think 
about ipods which need an eject). That's why they're named 
setup/teardown and are one the StorageAccess interface.


svn path=/trunk/KDE/kdelibs/; revision=674126