The old code resulted in two triggered(MouseButtons, KeyboardModifiers) signals if the middle mouse button was pressed. One from the direct invokation and one from the the call to QMenu::mouseReleaseEvent().
(And the triggered signal from QMenu caries the wrong mouse buttons.) 

With this commit, QMenu never sees the middle clicks. (Neither button down, nor button up.)
Which means triggered(bool) is never emited, just triggered(MouseButtons, KeyboarModifiers).
To me this seems sane. 

Because this afaik mirrows the toolbar actions behaviour and most of the widgets don't treat a
middle mouse click as a left mouse click.

(This fixes the middle mouse click in the bookmarks menu, opening the web site in both the
current and the new tab.)


svn path=/trunk/KDE/kdelibs/; revision=694011