the js keyboard event (mozilla extension) property event returns 0 if the key pressed cannot be expressed as an ascii code (apparently).
This behavior is the same as Firefox, /more/ similar to Opera (although
opera appears to have an implementation of "event" more buggy than
konqueror as a result of this patch). I figure firefox's implementation
is canonical as it's a gecko extension.

Still, there's a remaining bug in that you can't send "ctrl-something"
because Konqueror filters it away from the Part. I'm not going to fix that.


svn path=/branches/KDE/3.5/kdelibs/; revision=696502