Finally I have found some time to recompile KDE and test the feature again.
This is re-commit of implementation of switching language for individual applications.
Previous commit (r647712) was reverted as feature was not ported to KDE 4 then.
Now it is.

For language switching option see menu Help->Switch application language.

Feature freeze exception was granted for this feature. See:
http://lists.kde.org/?l=kde-devel&m=118096509723997&w=2

Thanks to release team for that :)

Also thanks to David Faure for review and providing valuable tips.

CCBUG: 24348
FEATURE: 24348
GUI:


svn path=/branches/KDE/3.5/kdelibs/; revision=700380