I think some lines in the kconfigdialogmanager.cpp file should be changed. The API doesn't change. I sent this already to the kde-core-devel mailing list and got an ok but nobody changed it. So I do it myself now.
The class QListBox is no longer part of Qt.

The parameter of the currentChanged signal in the QTabWidger class changed. It 
is now an integer and not a QWidget*

The name of the signal in KListWidget changed. KListWidget and QListWidget 
don't have the signal selectionChanged().

I think it should be itemSelectionChanged() instead.

svn path=/trunk/KDE/kdelibs/; revision=701793