Handle Pointer-to-Member-Functions (PMFs) as well.
Instead of trying to match the * in
    long long unsigned int* (* TestClass1::func_fptr())()

match the space that precedes it. It can't be the end of the function
name if there's a space. This now works for the PMFs:

long long unsigned int* (TestClass1::* TestClass1::func_pmf())()
long long unsigned int* (* (TestClass1::* TestClass1::func_uglypmf(long long unsigned int* (* (TestClass1::*)())()))())()

svn path=/trunk/KDE/kdelibs/; revision=702236