Use gmtoffset() rather than tm_gmtoff directly
Changed gmtoffset so that when tm_gmtoff is not available it uses the
extern variable timezone. timezone appears to be always available on
Unix type platforms because the KDE3 code uses it in the same way.
There is already code in place for windows, so that should be OK. In
addition the negative value of timezone is used, again because KDE3
did it that way.

svn path=/trunk/KDE/kdelibs/; revision=702282