This is an atomic commit. If needs to be reverted, can be done in one step, no more commits related to this one.
This adds the init() method to the KCModule that is called once that we know that the change() signal has been connected.

I didn't use QTimer::singleShot(0, ...) because it didn't work at least on my system. Hard coding a value on msec I thought it wasn't
very safe. For sanity when debugging and reading code, I added this method.

It also fixes the problematic plugin behavior. Is the only use case I found for this method to be added to the class.

Please, feel free to revert it if you think is necessary.

CCMAIL: kretz@kde.org

svn path=/trunk/KDE/kdelibs/; revision=705023