1) Convert Gregorian calendar to use base class methods instead of QDate methods 2) Add Unit Tests for converted methods 3) Fix bugs in base class methods found by unit tests :-) 4) Add lots of validity checks and tests (never trust input values) 5) Add more apidox
Converting Gregorian to the new base class is not strictly necessary, but
makes for a great test bed for the new methods as we can compare them directly
to the QDate results (except where QDate is broken!).  I'll probably switch 
some of these back again later for efficiency, but will leave for now as the
baseline tests.

This is Gregorian completed, as well as the base class except for weekNumber().

svn path=/trunk/KDE/kdelibs/; revision=708110