Rework how we sync textarea contents between the DOM and the Renderer, fixing multiple bugs (losing contents on display:none, the can't-do-anything-with-selection- in-wordpress bug).
This also improves compatibility with IE on handling of innerText somewhat, and that 
with other browsers in how we handle changes to the default value/child nodes.
Unfortunately, no one does it the same way, so I chose to follow Safari since its 
behavior makes the most sense to me. 

More specifically:
1. Simplify the syncing logic by making the renderer always be definitive when 
it exists.

2. Change how we  initialize from the defaultValue --- instead of doing it in 
the renderer's close (which is what causes the disappearing text bug), we 
update value to defaultValue when the children change. That also makes 
innerText work sensibly on textareas.

BUG: 132844
BUG: 120607
BUG: 141457


svn path=/trunk/KDE/kdelibs/; revision=712622