Sometimes we name our manpage docbook files as "man-foo.docbook", but the contents of that same file will say to generate a manpage called foo.1. (see <refentrytitle> in the docbook).
So, we change the kde4_create_manpage() macro to do a s+/man-+/+ when
creating the output file from the input file, i.e., man-foo.docbook
becomes foo.1.

The safer way would be to parse the docbook file, looking for
the <refentrytitle> value. Then create the output file from that.
Rather not, unless required later.


svn path=/trunk/KDE/kdelibs/; revision=717431