Fix toolbar style and size settings not being restored.
The settings were saved correctly, but were always being read from the
global config. The KConfigGroup with the global settings was declared
in an inner scope with the same name as the one holding the local
settings, causing it to be hidden, and the global one to be used in its
place.

svn path=/trunk/KDE/kdelibs/; revision=717442