Today's quizz: KConfigGroup(c,"").group() == ?
Answered by this unit test: it's "<default>", i.e. the toplevel entries in the config files.
I thought we had support for the empty group ("[]" in the config file) but obviously not.
Good I guess, one less thing to support. But then we could get rid of the "<default>" string altogether,
and use the empty string instead, no?
CCMAIL: kde-core-devel@kde.org

svn path=/trunk/KDE/kdelibs/; revision=721135