Implement icon fallbacks, as specified by the fd.o icon naming specification. The only slight difference is that mimetype icons are not special cased and therefore also look for fallbacks.
This means that you (as an application developer) can now specify
specific icons like go-next-year-korganizer, and if this icon
doesn't exist it will retrieve go-next-year instead, and if this
doesn't exist as well, you'll get go-next (which is specified
in the icon naming spec and therefore can be expected to exist).

If you ever needed a good reason to name your icons in a
proper way, this is it.

p0nies?

CCMAIL: kde-artists@kde.org
CCMAIL: aseigo@kde.org

svn path=/trunk/KDE/kdelibs/; revision=721319