KService: support for storing Actions defined in .desktop files into ksycoca4.
While writing the unit test for this, I noticed that the binary representation (like, in ksycoca)
of a single KService was 3700 bytes. This is because KConfigGroup::entryMap returns all translated
entries. After filtering those out, the KService is down to 755 bytes; and the whole ksycoca4 went
from 9.0M to 1.5M! Seems to be an unexplained KConfig behavior change, which should be fixed, though.

svn path=/trunk/KDE/kdelibs/; revision=725525