improve the speed of KIconLoader::loadIcon and iconPath dramatically. for cached icons, this takes us from 22k icons/s to somewhere around 118k.
windows people: you still have the bane of QDir in this hot path, but i don't know what to replace it with that will work on windows. QDir is just too expensive to use in this hot path just to check for a relative or absolute path. if you know of a clever way to do this without resorting to QDir, please make the change in the appropriate ifdef areas. otherwise you'll only get about half the benefit here.

bug 151874: i still don't know if you're having issues with the caching of icons. please test with this rev and let me know if this helps at all.

CCMAIL:kde-window@kde.org
BUGS:151874

svn path=/trunk/KDE/kdelibs/; revision=733164