Get rid of the arrow-{up,down,left,right} usage in KHTML by using the style arrows for drawing the mouse scroll indicator.
However, this still makes the indicator not appear -
I can see it shining through (occasional one-pixel shifts
make it look like transparent glass) but it's not drawn
with a visible color obviously. No idea which brush,
composition mode or whatever is to blame for this.

Also, use two more appropriate icon names.

svn path=/trunk/KDE/kdelibs/; revision=735559