prepare getting rid of the nonCSSHints flag, for complete inline  style correctness.
We can't keep a mixed inline style/presentational style list has
retrieving the style() of a node from the DOM should never expose the
presentational styles, only the inline ones.

In order to avoid making ElementImpl grow, we'll host the pointers
to inline style and presentational style within a struct created on-demand,
unioned with the inline styles pointer.

svn path=/trunk/KDE/kdelibs/; revision=736405