We've always gotten some trouble from the fact that KHTMLFactory has a dual purpose; being the KParts::Factory (when khtmlpart is dlopened, e.g. in konqueror), and being the repository for a bunch of global things needed by khtml.
I've separated those two things, creating a KHTMLGlobal for the global stuff and leaving
the KHTMLFactory to being the kparts factory only. It solves some life-cycle issues
and allows to remove the ugly bool clone from the constructor. I also improved
the refcounting mechanism in KHTMLGlobal to make it debuggable.

This doesn't fix the KHTMLFactory assert though; this only showed that the problem
lies in DocumentImpl refcounting, not in KHTMLFactory itself...

OK'ed by carewolf.

svn path=/trunk/KDE/kdelibs/; revision=744435