Fixes the problem of going to "kcmshell4 fonts", change the toolbar font and not being changed dynamically the font on those running applications. You have to restart them to see the change happen. The KToolBar was setting its own font at the constructor, and that means that when we later changed the QApplication font through QApplication::setFont the toolbar wasn't updated, so it remained with the old font.
The fix is to remove the code at the KToolBar private constructor, as long as this font is also set from the KGlobalSettings object being created when the instance of a 
program is created. This will set the toolbars fonts, and this method will also be called when it needs to be updated because a signal was received for that reason.

CCMAIL: kde-core-devel@kde.org

svn path=/trunk/KDE/kdelibs/; revision=746723