Fix an issue in the canvas arc() implementation.
Adjusting the end point by inserting a MoveTo element when the sweep length
is greater than 360 degrees is not sufficient, since it won't always produce
correct results if additional elements are added to the path before it is
stroked or filled. Call QPainterPath::arcTo() twice instead, to make the arc
continue to the intended end point.

svn path=/trunk/KDE/kdelibs/; revision=753535