Implement support for repeat-x, repeat-y and no-repeat in CanvasPatternImpl.
This is done by calculating the area that should be filled with the pattern
given the repeat setting, and using this area as a clip rect when drawing
the path. Add a drawPath() helper function in CanvasContext2DImpl that takes
care of applying and restoring the clip region, and also calls
drawPathWithShadow() when needed.

svn path=/trunk/KDE/kdelibs/; revision=753567