Allow padding of form elements in the case of border-box, which causes the form element widget to get bigger but its active region to stay the same size.
This is sometimes used to put little icons on lineedits.

In the case of the content-box model, the padding goes around the form element,
as it is today (also the default in non-quirky mode).

There's still a strange bug with how the scrollbar on scrollable widgets
doesn't have any hover effects and it has the "text input" cursor. But that's not
a regression and is tolerable, so we decided to accept it until we have a chance
to clean up render_replaced.

Reviewed repeatedly by Germain.

BUG: 109068

svn path=/trunk/KDE/kdelibs/; revision=755231