Fix occasional blank KIO progress windows that crash when you interact with them
The problem was that jobs were not unregistered, and so if you happened to get
a job allocated where a previous job had been, the old job was still registered
and confused everything. Fix this by unregistering in the tracker widget's destructor,
but make sure not to delete the widget twice.

CCBUG: 151759

svn path=/trunk/KDE/kdelibs/; revision=758150