Support the HTML tag <bdo> (bidirectional override). As a result, support CSS unicode-bidi: override.
First, don't change two CSS properties in html_elementimpl.cpp as it may
unexpectedly overwrite a previously set property. This allows us to add the
"selector that seems to be still broken" in khtml4.css.

Then make Qt render override text in RTL, even if it's Latin script.

Add a kate variable line because it will make me happy.

Also, fix a bug where it would render <bdo> as not rtl the first time you try
to do bidi stuff in konq (per process). This is because some global (!)
variables were being used unintialized.

There are still some bugs remaining (non-regression) related to the
application of the override. I have some testcases for this but it
requires a sacrifice to the bidi.cpp gods.

BUG:150515


svn path=/trunk/KDE/kdelibs/; revision=759219