automatically merged revision 760045: Fix some of the major performance problems this causes in textareas in KHTML by not computing spelling correction suggestions when none are needed, as they are extremely expensive to compute... Doesn't help kmail since it does connect the signal... A better option would be to do the computation on demand.
Also fix a clear logic error, and making word counting make sense some of the time and not
none of the time. 

This class needs heavy work, however. It rehighlights the entire document when one presses 
enter. The word counting is still bogus between rehighlights. It hardcodes black 
when switching the error off... IOW, it needs love from someone with good understanding of
QSyntaxHighlighter

svn path=/trunk/KDE/kdelibs/; revision=760455