automatically merged revision 761089: Move object loading to the DOM from the renderer, fixing the long-standing major bug that hidden iframes couldn't be interacted with.
The basic overview here is that a new DOM base class, HTMLPartContainerElementImpl
is created, that manages the interaction with KHTMLPart when it
comes to loading child parts. KHTMLPart now keeps tracks of those
per ChildFrame (along the way removing the confusion of having 2 m_frames
in close quarters talking about different things); and the DOM
objects for iframe/object/embed/frame/etc. request the loading themselves
as needed. The renderer "just" displays the part widget set from the DOM

BUG: 71809
BUG: 150240
(and probably others)

svn path=/trunk/KDE/kdelibs/; revision=761094