backport SVN commit 770700 by mkretz:
This code is _very_ old (it's a leftover from the first KDEMM prototype for
KDE3) and I can't find a reason why it's still needed. To the contrary I think
it made VideoPlayer behave different than described in the docs. (I.e. it would
start playing even if you called only load())

svn path=/branches/KDE/4.0/kdelibs/; revision=770702