(windows-only)
A workaround for OSes where DebugBreak() does not work in administrative mode (actually Vista with msvc 2k5) - display a native message box so developer can attach the debugger to the KIO slave process and click OK.

The patch allows to use
 set KDE_SLAVE_DEBUG_POPUP=protocol
instead of
 set KDE_SLAVE_DEBUG_WAIT=protocol

to get the popup for any version of msvc and any windows version.
As no DebugBreak() is the executed in "KDE_SLAVE_DEBUG_POPUP" mode, depending on the actual needs, user can just click OK without attaching, and continue using the KIO slave.

You can find more info on the source of the problem in "(Vista only)" note of [1]. 
I have also updated [1] for the new variable.

[1] http://techbase.kde.org/index.php?title=Development/Tutorials/Debugging/Debugging_on_MS_Windows#Debugging_kioslaves


svn path=/trunk/KDE/kdelibs/; revision=771942