(finally) backport: SVN commit 764908 by braxton:
This should fix the problem with saving settings on app exit.
The problem was I was unintentionally creating extra references to the
global component-data by leaking the lockfile.
This caused the global component-data d-tor to never be called therefore
the d-tor for the global config-file was never called so it didn't sync
on exit.


svn path=/branches/KDE/4.0/kdelibs/; revision=778323