Honor KActions "globalShortcutAllowed" and "shortcutConfigurable".
If the shortcut is advertised as not configurable by the developer, do not allow editing at all.
If the shortcut is not advertised as globalShortcutAllowed by the developer, do not allow setting of a global shortcut.

That breaks a lot of global shortcuts. They are opt in. I will fix them.

And it told me that Kate advertises it's "About Editor Component" as a global shortcut. Thanks for the testcase. But where is the sense in that?


svn path=/trunk/KDE/kdelibs/; revision=780104