automatically merged revision 783804: Tighten our paint prevention timings to provide more responsiveness while keeping smooth page transitions.
Proper balance is not easy to achieve, as preventing the
"white flash" is just part of the equation.
Layout stability, allowing early readability, is another important
term - as is bandwidth. Engines with aggressive timings would get an
unpleasnt, jerky feeling we want to avoid.

We'll use subtly progressive timings to get a good compromise.

svn path=/branches/KDE/4.0/kdelibs/; revision=784718