(reviewed, Windows only) As discussed before in the thread [1], there are problems with running KUniqueApplications on Windows because previous dbus service has not been unregistered, resulting in "KUniqueApplication: Can't setup D-Bus service. Probably already running." and forced app quit.
Now e.g. KMail runs just fine every time I executed it, 
what was not the case before (when we needed to kill klauncher and/org 
dbus-daemon - rather not an option for end users we're starting to demo 
KDE apps on Windows to :)).

There are sane arguments that certain fixes should be performed 
at the dbus level, but for now we have another month after 
the initial discussion and there is nothing better that this KDElibs-level fix. 
Thus, this change as temporary.

[1] http://lists.kde.org/?t=120294760700003&r=1&w=2 


svn path=/trunk/KDE/kdelibs/; revision=786562