CSS 3 - Media Queries
cf. http://www.w3.org/TR/css3-mediaqueries/

This is a full implementation supporting all
media features, for both visual and paged media types we support.

It is based on the partial implementation found in WebCore.
With some refactoring work, fixes, and implementation of
the missing bits.

This module allows web content writers to query the capabilities of
the client's output device (be it a screen, a printer, etc.)
and automatically select CSS styles that are best suited to it.

svn path=/trunk/KDE/kdelibs/; revision=788248