Make 3rd-party-theme fallbacks work correctly, part 2: Fix the theme fallback order in KIconLoader in order to a) include Oxygen as fallback even if it's not included    in the theme's "Inherits" property, and b) make sure that "hicolor" is always included and    always goes last in the theme hierarchy,    as defined by the icon theme specification.
(a) is what makes you get rid of the numerous "unknown" icons
if you use a theme that doesn't inherit Oxygen directly.
Enjoy running KDE with gnome-icon-theme & Co.! :P

Part 1 (reviving the default.kde symlink, r794564) wasn't
actually needed in order to make this work, but is still
a good thing to have, so I won't revert it.

svn path=/trunk/KDE/kdelibs/; revision=794709