backports for kfile
r794821, r795837:
A set of fixes for improving experience of file dialogs, mostly on Windows
+ some fixes for handling urls

KFileWidget:
- check whether a string contains protocol part more carefully because c:/ is
an absolute path, not a protocol
   (static containsProtocolSection()), so we assume the protocol has at least
two characters before ":/"
- enclose locationEdit->currentText() with QDir::fromNativeSeparators() and
add .trimmed() to cover more cases in conditions
- slotOk(): add code for handling accepting using OK button when directory
path is entered and directories are not acceptable result of the file widget:
cd to that directory then; in particuar it is now possible to enter C:\ or /
and press OK to change to the root dir on windows (or linux, respectively)


r795936:
Fix changing the current dir using the 'recent url' combo box.
Upon accepting new url in the 'directory' url combo box, append '/' if needed: the combo does not add it, but tokenize() expects it because uses KUrl::setFileName(), which would silently remove the last segment of the path.

Without the fix, if you select /home/js/foo/bar from the url combo box, and click on file.txt, the resulting url would be /home/js/foo/file.txt, not the expected /home/js/foo/bar/file.txt.

Another, and hopefully the last place where we have to care about trailing slashes for dirs, is kdiroperator.

(reviewed)


svn path=/branches/KDE/4.0/kdelibs/; revision=795939