Preloading of network resources via a side-tokenizer.
based on patch by Antti Koivisto,
ported and reworked so that it is simpler and more efficient.

What it does is, while the main tokenizer is blocked because of a
script that's not yet arrived - scan ahead the content we have already
received in order to locate and start requesting other resources.

This cuts the loading time in half on many high profile pages!

svn path=/trunk/KDE/kdelibs/; revision=798245