-use export_library_dependencies() into an extra file, which is included then, so the position of the set()-commands can be manipulated, otherwise they are always completely at the end of the file
-add option KDE4_ENABLE_EXPERIMENTAL_LIB_EXPORT, which enables an experimental new way
how the library dependencies are exported.
Instead of regular cmake export_library_dependencies() this is now done "manually" using the
LINK_INTERFACE_LIBS target property. This property holds a list of libraries which stuff that
links against this library will also be linked against.
Once we require cmake 2.6.x (KDE 4.2 ?) this can be used as is for using the new-style
cmake export functions

Alex


svn path=/trunk/KDE/kdelibs/; revision=810531