Prevent a crash when looking up U+FDD0 and U+FDD1.
Qt internally uses U+FDD0 and U+FDD1 to mark the beginning and the end of frames.
They should be seen as non-printable characters, as trying to display them leads
to a crash caused by a Qt "noBlockInString" assertion.

Thanks to http://xkcd.com/380/ for pointing me to the problem. :) (Look at the tooltip.)


svn path=/trunk/KDE/kdelibs/; revision=814947