This fixes a bit tricky bug. The default places are not "system" ones. I cannot find that on the spec, and that leads to a problem following the next steps:
- Open kwrite, the open/save dialog. You have the default places.
- Add a new place to them (it will be added to
  ~/.local/share/user-places.xbel) - only that new place, not Home, Root
or others, only that one.
- Remove ~/.kde or ~/.kde4, your local KDE 4 folder information.
- Open kwrite, the open/save dialog now only contains your own places.

This is a consequence of adding them as system ones, which makes sense
for devices, but not for this places. Now they are all written to
~/.local/share/user-places.xbel.

Not having suddenly Home, Root & friends on the places view is bad, and
you also get a very ugly "Custom Path > Home > Foo" on the top of the
dialog, not even "Root > Home > Foo", since not even root exists on the
places view.

Thanks to annma for spotting the bug and for ana` for helping out me for
reproducing it.

CCMAIL: annma@kde.org
CCMAIL: nf2@scheinwelt.at

svn path=/trunk/KDE/kdelibs/; revision=820333