Add an aggressive cleanup to the beginning of kdeinit, just in case it gets file descriptors leaked from its parent and ancestors.
kdeinit4 is now clean for its own file descriptors.

It's not clean if the libraries it uses leak. However, since it
doesn't initialise any of those libraries, leakages can only occur as
a result of global constructors being run (and those are evil on their
own).

A shell inside konsole will show either 0 leakages, or 4 leakages if
glib is in use and its version doesn't have the following bug closed:

http://bugzilla.gnome.org/show_bug.cgi?id=538119

svn path=/trunk/KDE/kdelibs/; revision=821132