-also disable _KDE4_TARGET_LINK_INTERFACE_LIBRARIES() and use the renamed LINK_INTERFACE_LIBRARIES target property instead: set_target_properties(foo PROPERTIES                           ${KDE4_DISABLE_PROPERTY_}LINK_INTERFACE_LIBRARIES  "kdeui;kdecore"                      )
While this doesn't look very nice, it is completely forward and backward compatible (i.e. cmake 2.4 doesn't complain,
cmake 2.6 works, both versions work as soon as the reduced link interface is the default (soon), if somebody 
uses this with kdelibs which don't have that anymore he will get the reduced interface instead of errors, so all 
should be ok)

This was discussed on kde-buildsystem end of June.

Alex

CCMAIL: mueller@kde.org


svn path=/branches/KDE/4.1/kdelibs/; revision=836133