Provide a signal aboutToStealShortcut from KKeySequenceWidget and ShortcutEditWidget(forwarded).
Do not call KKeySequenceWidget::setKeySequence from ShortcutEditWidget if
not necessary. That method has some side effects.

Document a bug in applyStealShortcut. It's not a good idea to save here.
The bug should only bite people trying to implement their own version of
KShortcutsEditorWidget. It brings the *ui.rc in a bad state if you are
unlucky :-( .

svn path=/trunk/KDE/kdelibs/; revision=855019