Merge from kdelibs-4.0.83 branch of:
r836515 | tilladam | 2008-07-22 14:48:23 +0200 (Tue, 22 Jul 2008) | 1 line

When using the QFileSystemWatcher backend of KDirWatch, trust it to correctly 
notify us when a file or diretory changes, instead of second guessing it based
on mtime. Adding a file to a directory does not always update the directory
mtime, for example, which makes this check fail in some situations.
kolab/issue2880


svn path=/branches/kdepim/enterprise4/kdelibs-4.1-branch/; revision=856398