Use a simple search index for KCharSelect and implement instant search
This also makes searching in character details possible. The indexing
of Unihan data is currently disabled, since it increases the size of
the index a lot. Probably an option should be introduced for this.

The index is created on runtime in the background (via QtConcurrent),
so the only time the UI could freeze a few milliseconds is if the user
tries to search before the index is ready.

FEATURE: 167423


svn path=/trunk/KDE/kdelibs/; revision=856440